feat: (Search) - refactor with a SearchUtils
